Add checkboxes to the zone panel
Zone panel contains the list of Montreal's borough with
an associated checkbox. When a checkbox is checked (or unchecked)
the addBorough (or removeBorough) action is fired with the proper
borough's name.

Note: when other zone type will be supported, borough selection
will be moved to a dedicated panel called BoroughPanel.